debugging

+++ events/messageReactionAdd.js
@@ -4,7 +4,7 @@ module.exports = {
   },
   exec: async (client, reaction, user) => {
     user = await client.database.users.findOne({discord_id: user.id})
-    if (reaction.emoji.name == '❌' && user.auth_level >= 9) {
+    if (reaction.emoji.name == '❌' && user.auth_level >= 9 && reaction.message.deletable == true) {
       reaction.message.channel.send('\`\`deleting\`\`').then(async (message) => {
          await client.globals.sleep(2000)
          message.delete()
